pyproject.toml
[tool.poetry]
name = "MyAutoML"
version = "0.1.0"
description = ""
authors = ["nomnomnonono <nozomu.koujiguchi1204@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.0"
python-dotenv = "^1.0.0"
mlflow = "^2.6.0"
streamlit = "^1.26.0"
google-cloud-storage = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
